requirements.txt
streamlit
openai
pandas
pydantic
python-dotenv
